WE HAVE LIFTOFF!!!!!!!!!!!!!!!!!!!!!!!! Numerous OS / Platform backend developments, some graphics bug fixes and caching, etc... Today is a good day!

+++ src/graphics/graphics_device_managed_resources.h
@@ -3,7 +3,7 @@
 	typedef struct GFX_##mgp_name##Key GFX_##mgp_name##Key;				\
 	struct GFX_##mgp_name##Key											\
 	{																	\
-		u32 value;														\
+		u64 value;														\
 	};																	\
 	internal GFX_##mgp_name##Key GFX_##mgp_name##KeyNull(void);			\
 	internal b32 GFX_##mgp_name##KeyIsNull(GFX_##mgp_name##Key key);	\
